Further edge-of-world fixes + chat rendering issue fix + entity management fix

+++ perovskite_client/src/vulkan/far_geometry.rs
@@ -1,7 +1,7 @@
 use std::{collections::hash_map::Entry, sync::Arc};
 
 use anyhow::{Context, Result};
-use cgmath::{vec4, ElementWise, InnerSpace, Matrix4, Vector3, Vector4};
+use cgmath::{vec4, ElementWise, InnerSpace, Matrix4, Vector3};
 use itertools::Itertools;
 use perovskite_core::{
     coordinates::ChunkCoordinate,
@@ -170,8 +170,8 @@ impl ClientFarSheet {
         );
 
         let chunk_corners = control.lattice_corners_world_space().map(|corner| {
-            let x = (corner.x.clamp(i32::MIN as f64, i32::MAX as f64) as i32).div_euclid(16);
-            let z = (corner.y.clamp(i32::MIN as f64, i32::MAX as f64) as i32).div_euclid(16);
+            let x = (corner.x.clamp(i32::MIN as f64, i32::MAX as f64) as i64).div_euclid(16) as i32;
+            let z = (corner.y.clamp(i32::MIN as f64, i32::MAX as f64) as i64).div_euclid(16) as i32;
             (x, z)
         });
